Renamed DebugMode.java to LogLevel.java [=] RestAssuredRequest.java and RestAssuredResponse.java [=] LogLevel.java from DebugMode.java [=] core.properties and example-api.properties
