+++ src/main/resources/core.properties
@@ -1,1 +1,1 @@
-#debugMode=prod
+#logLevel=prod
